used size_t as type for buffer lengths

+++ src/mumble/pluginAPI/Plugin.h
@@ -9,6 +9,7 @@
 
 #include "PluginComponents.h"
 #include <stdint.h>
+#include <stddef.h>
 
 #ifdef WIN32
 #define PLUGIN_EXPORT __declspec(dllexport)
@@ -44,13 +45,13 @@ extern "C" {
 	/// @param[out] nameBuffer The buffer into which the name of the plugin shall be copied by this function
 	/// @param bufferSize The size of the buffer - The length of the name of the plugin (including the terminating zero-byte) must not exceed
 	/// 	this length
-	PLUGIN_EXPORT void getName(char *nameBuffer, int32_t bufferSize);
+	PLUGIN_EXPORT void getName(char *nameBuffer, size_t bufferSize);
 
 	/// Gets the plugin's version in a display-ready representation.
 	///
 	/// @param[out] versionBuffer The buffer into which the version string shall be copied by this function
 	/// @param bufferSize The size of the buffer - The length of the version (including terminating zero-byte) must not exceed this length
-	PLUGIN_EXPORT void getDisplayVersion(char *versionBuffer, int32_t bufferSize);
+	PLUGIN_EXPORT void getDisplayVersion(char *versionBuffer, size_t bufferSize);
 
 	/// Gets the Version of this plugin
 	///
@@ -68,14 +69,14 @@ extern "C" {
 	/// @param[out] authorBuffer The buffer into which the author name shall be copied by this function
 	/// @param bufferSize The size of the buffer - The length of the author name (including the terminating zero-byte) must not exceed
 	/// 	this length.
-	PLUGIN_EXPORT void getAuthor(char *authorBuffer, int32_t bufferSize);
+	PLUGIN_EXPORT void getAuthor(char *authorBuffer, size_t bufferSize);
 
 	/// Gets the description of the plugin
 	///
 	/// @param[out] descriptionBuffer The buffer into which the description shall be copied by this function
 	/// @param bufferSize The siue of the buffer - The length of the description (including the terminating zero-byte) must not exceed
 	/// 	this length.
-	PLUGIN_EXPORT void getDescription(char *descriptionBuffer, int32_t bufferSize);
+	PLUGIN_EXPORT void getDescription(char *descriptionBuffer, size_t bufferSize);
 
 	/// Provides the MumbleFunctions struct to the plugin. This struct contains function pointers that can be used
 	/// to interact with the Mumble client. It is up to the plugin to store this struct somewhere if it wants to make use
